feat: delection section separated to component
